Move Common files out of DecodableReader
