+++ src/BloomBrowserUI/bookEdit/toolbox/readers/decodableReader/decodableReaderToolboxPanel.ts
@@ -1,10 +1,10 @@
 ﻿/// <reference path="../../toolbox.ts" />
-/// <reference path="./directoryWatcher.ts" />
-/// <reference path="./readerToolsModel.ts" />
+/// <reference path="../directoryWatcher.ts" />
+/// <reference path="../readerToolsModel.ts" />
 
-import {DirectoryWatcher} from "./directoryWatcher";
-import {DRTState, ReaderToolsModel, MarkupType} from "./readerToolsModel";
-import {beginInitializeDecodableReaderTool} from "./readerTools";
+import {DirectoryWatcher} from "../directoryWatcher";
+import {DRTState, ReaderToolsModel, MarkupType} from "../readerToolsModel";
+import {beginInitializeDecodableReaderTool} from "../readerTools";
 import {ITabModel} from "../../toolbox";
 import {ToolBox} from "../../toolbox";
 import theOneLocalizationManager from '../../../../lib/localizationManager/localizationManager';
