+++ src/BloomBrowserUI/bookEdit/toolbox/readers/directoryWatcher.ts
@@ -1,4 +1,4 @@
-/// <reference path="../../../../typings/jquery/jquery.d.ts" />
+/// <reference path="../../../typings/jquery/jquery.d.ts" />
 
 /**
  * Implements a simple directory watcher in javascript. The localhost is listening for requests that begin with this
