+++ src/BloomBrowserUI/bookEdit/toolbox/readers/readerToolsSpec.ts
@@ -1,6 +1,6 @@
-import {theOneLanguageDataInstance, ResetLanguageDataInstance}  from '../libSynphony/synphony_lib';
+import {theOneLanguageDataInstance, ResetLanguageDataInstance}  from './libSynphony/synphony_lib';
 import {ReaderToolsModel} from './readerToolsModel';
-import ReadersSynphonyWrapper from '../ReadersSynphonyWrapper';
+import ReadersSynphonyWrapper from './ReadersSynphonyWrapper';
 
 describe("Bloom Edit Controls tests", function() {
     var classValues;
