+++ src/BloomBrowserUI/bookEdit/toolbox/toolboxBootstrap.ts
@@ -1,7 +1,7 @@
 /// <reference path="../../typings/jquery/jquery.d.ts" />
 import * as $ from 'jquery';
 import {restoreToolboxSettings, showOrHidePanel_click, removeToolboxMarkup} from './toolbox';
-import {ReaderToolsModel} from './readers/decodableReader/readerToolsModel'
+import {ReaderToolsModel} from './readers/readerToolsModel'
 import TalkingBookModel from './talkingBook/talkingBook';
 import LeveledReaderToolboxPanel from './readers/leveledReader/leveledReaderToolboxPanel';
 import {handleBookSettingCheckboxClick} from './bookSettings/bookSettings';
@@ -10,8 +10,8 @@ import {handleBookSettingCheckboxClick} from './bookSettings/bookSettings';
 // reviewslog: is this actually needed? Could these be be directly imported where they are used?
 export {showOrHidePanel_click};
 export {removeToolboxMarkup};
-export {showSetupDialog, initializeReaderSetupDialog, closeSetupDialog} from './readers/decodableReader/decodableReaderToolboxPanel'
-export {addWordListChangedListener, beginSaveChangedSettings} from './readers/decodableReader/readerTools';
+export {showSetupDialog, initializeReaderSetupDialog, closeSetupDialog} from './readers/decodableReader/decodableReaderToolboxPanel';
+export {addWordListChangedListener, beginSaveChangedSettings} from './readers/readerTools';
 export {loadLongpressInstructions} from '../js/bloomEditing';
 export {TalkingBookModel}; // one function is called by CSharp; also, exporting something from it gets it included in the bundle.
 export {LeveledReaderToolboxPanel}; // just to make sure it gets included in the bundle (and adds an instance of itself to the collection in toolbox.ts)
